requirements.txt
aiofiles==23.1.0
aiogram==3.0.0rc1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
frozenlist==1.4.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
l10n==0.1.5
magic-filter==1.0.11
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
openai==0.27.8
polib==1.2.0
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
termcolor==2.3.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
